feat : MemberTestFixture 생성
